Fixing and cleaning up
