+++ src/lib.rs
@@ -6,5 +6,6 @@ pub mod http_api;
 pub mod http_service;
 pub mod udp_common;
 pub mod udp_service;
-pub mod tracker;
 pub mod databases;
+pub mod tracker;
+pub mod tracker_objects;
