+++ src/main.rs
@@ -9,13 +9,15 @@ use std::net::SocketAddr;
 use std::process::exit;
 use std::time::Duration;
 use tokio::time::timeout;
+
 use torrust_axum::common::{tcp_check_host_and_port_used, udp_check_host_and_port_used};
 use torrust_axum::config::{Configuration, DatabaseStructureConfig};
 use torrust_axum::databases::DatabaseDrivers;
 use torrust_axum::http_api::{http_api, https_api};
 use torrust_axum::http_service::{http_service, https_service};
 use torrust_axum::logging::setup_logging;
-use torrust_axum::tracker::{StatsEvent, TorrentTracker};
+use torrust_axum::tracker::TorrentTracker;
+use torrust_axum::tracker_objects::stats::StatsEvent;
 use torrust_axum::udp_service::udp_service;
 
 #[global_allocator]
@@ -177,7 +179,7 @@ async fn main() -> std::io::Result<()>
                 tracker_send.add_shadow(*info_hash, *completed).await;
                 tracker_receive.remove_torrent(*info_hash, false).await;
             }
-            start = start + amount;
+            start += amount;
         }
 
         let _ = tracker_send.save_torrents().await;
