+++ src/tracker.rs
@@ -1,188 +1,14 @@
-use chrono::{TimeZone, Utc};
-use log::{error, info};
+use chrono::Utc;
 use scc::ebr::Arc;
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
 use std::collections::{BTreeMap, HashMap};
-use std::ops::Add;
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
-use async_std::future::timeout;
-use tokio::sync::{Mutex, RwLock};
+use tokio::sync::Mutex;
 
-use crate::common::{InfoHash, NumberOfBytes, PeerId, TorrentPeer};
+use crate::common::{InfoHash, PeerId, TorrentPeer};
 use crate::config::Configuration;
 use crate::databases::DatabaseConnector;
-
-#[derive(Serialize, Deserialize, Clone, Copy)]
-pub enum StatsEvent {
-    Torrents,
-    TorrentsUpdates,
-    TorrentsShadow,
-    Users,
-    UsersUpdates,
-    UsersShadow,
-    TimestampSave,
-    TimestampTimeout,
-    TimestampConsole,
-    TimestampKeysTimeout,
-    MaintenanceMode,
-    Seeds,
-    Peers,
-    Completed,
-    Whitelist,
-    Blacklist,
-    Key,
-    Tcp4ConnectionsHandled,
-    Tcp4ApiHandled,
-    Tcp4AnnouncesHandled,
-    Tcp4ScrapesHandled,
-    Tcp6ConnectionsHandled,
-    Tcp6ApiHandled,
-    Tcp6AnnouncesHandled,
-    Tcp6ScrapesHandled,
-    Udp4ConnectionsHandled,
-    Udp4AnnouncesHandled,
-    Udp4ScrapesHandled,
-    Udp6ConnectionsHandled,
-    Udp6AnnouncesHandled,
-    Udp6ScrapesHandled,
-}
-
-#[derive(Serialize, Deserialize, Clone, Copy)]
-pub struct Stats {
-    pub started: i64,
-    pub timestamp_run_save: i64,
-    pub timestamp_run_timeout: i64,
-    pub timestamp_run_console: i64,
-    pub timestamp_run_keys_timeout: i64,
-    pub torrents: i64,
-    pub torrents_updates: i64,
-    pub torrents_shadow: i64,
-    pub users: i64,
-    pub users_updates: i64,
-    pub users_shadow: i64,
-    pub maintenance_mode: i64,
-    pub seeds: i64,
-    pub peers: i64,
-    pub completed: i64,
-    pub whitelist_enabled: bool,
-    pub whitelist: i64,
-    pub blacklist_enabled: bool,
-    pub blacklist: i64,
-    pub keys_enabled: bool,
-    pub keys: i64,
-    pub tcp4_connections_handled: i64,
-    pub tcp4_api_handled: i64,
-    pub tcp4_announces_handled: i64,
-    pub tcp4_scrapes_handled: i64,
-    pub tcp6_connections_handled: i64,
-    pub tcp6_api_handled: i64,
-    pub tcp6_announces_handled: i64,
-    pub tcp6_scrapes_handled: i64,
-    pub udp4_connections_handled: i64,
-    pub udp4_announces_handled: i64,
-    pub udp4_scrapes_handled: i64,
-    pub udp6_connections_handled: i64,
-    pub udp6_announces_handled: i64,
-    pub udp6_scrapes_handled: i64,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct TorrentEntryItem {
-    pub completed: i64,
-    pub seeders: i64,
-    pub leechers: i64,
-}
-
-impl TorrentEntryItem {
-    pub fn new() -> TorrentEntryItem {
-        TorrentEntryItem {
-            completed: 0,
-            seeders: 0,
-            leechers: 0,
-        }
-    }
-}
-
-impl Default for TorrentEntryItem {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct TorrentEntry {
-    #[serde(skip)]
-    pub peers: BTreeMap<PeerId, TorrentPeer>,
-    pub completed: i64,
-    pub seeders: i64,
-    pub leechers: i64,
-}
-
-impl TorrentEntry {
-    pub fn new() -> TorrentEntry {
-        TorrentEntry {
-            peers: BTreeMap::new(),
-            completed: 0,
-            seeders: 0,
-            leechers: 0,
-        }
-    }
-}
-
-impl Default for TorrentEntry {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct UserEntryItem {
-    pub uuid: String,
-    pub key: String,
-    pub uploaded: i64,
-    pub downloaded: i64,
-    pub completed: i64,
-    pub updated: i64,
-    pub active: i64,
-}
-
-impl UserEntryItem {
-    pub fn new() -> UserEntryItem {
-        UserEntryItem {
-            uuid: "".to_string(),
-            key: "".to_string(),
-            uploaded: 0,
-            downloaded: 0,
-            completed: 0,
-            updated: 0,
-            active: 0,
-        }
-    }
-}
-
-impl Default for UserEntryItem {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone)]
-pub struct GetTorrentsApi {
-    pub info_hash: String,
-    pub completed: i64,
-    pub seeders: i64,
-    pub leechers: i64,
-}
-
-#[derive(Serialize, Deserialize, Clone)]
-pub struct GetTorrentApi {
-    pub info_hash: String,
-    pub completed: i64,
-    pub seeders: i64,
-    pub leechers: i64,
-    pub peers: Vec<Value>,
-}
+use crate::tracker_objects::stats::Stats;
+use crate::tracker_objects::torrents::TorrentEntryItem;
+use crate::tracker_objects::users::UserEntryItem;
 
 pub struct TorrentTracker {
     pub config: Arc<Configuration>,
@@ -252,138 +78,6 @@ impl TorrentTracker {
         }
     }
 
-    /* === Statistics === */
-    pub async fn get_stats(&self) -> Result<Stats, ()>
-    {
-        match timeout(Duration::from_secs(30), async move {
-            let stats_arc = self.stats.clone();
-            let stats_lock = stats_arc.lock().await;
-            let stats = stats_lock.clone();
-            drop(stats_lock);
-            stats
-        }).await {
-            Ok(data) => { Ok(data) }
-            Err(_) => { error!("[GET_STATS] Read Lock (stats) request timed out!"); Err(()) }
-        }
-    }
-
-    pub async fn update_stats(&self, event: StatsEvent, value: i64) -> Stats
-    {
-        let mut count = 1;
-        loop {
-            match timeout(Duration::from_secs(30), async move {
-                let stats_arc = self.stats.clone();
-                let mut stats_lock = stats_arc.lock().await;
-                match event {
-                    StatsEvent::Torrents => { stats_lock.torrents += value; }
-                    StatsEvent::TorrentsUpdates => { stats_lock.torrents_updates += value; }
-                    StatsEvent::TorrentsShadow => { stats_lock.torrents_shadow += value; }
-                    StatsEvent::Users => { stats_lock.users += value; }
-                    StatsEvent::UsersUpdates => { stats_lock.users_updates += value; }
-                    StatsEvent::UsersShadow => { stats_lock.users_shadow += value; }
-                    StatsEvent::TimestampSave => { stats_lock.timestamp_run_save += value; }
-                    StatsEvent::TimestampTimeout => { stats_lock.timestamp_run_timeout += value; }
-                    StatsEvent::TimestampConsole => { stats_lock.timestamp_run_console += value; }
-                    StatsEvent::TimestampKeysTimeout => { stats_lock.timestamp_run_keys_timeout += value; }
-                    StatsEvent::MaintenanceMode => { stats_lock.maintenance_mode += value; }
-                    StatsEvent::Seeds => { stats_lock.seeds += value; }
-                    StatsEvent::Peers => { stats_lock.peers += value; }
-                    StatsEvent::Completed => { stats_lock.completed += value; }
-                    StatsEvent::Whitelist => { stats_lock.whitelist += value; }
-                    StatsEvent::Blacklist => { stats_lock.blacklist += value; }
-                    StatsEvent::Key => { stats_lock.keys += value; }
-                    StatsEvent::Tcp4ConnectionsHandled => { stats_lock.tcp4_connections_handled += value; }
-                    StatsEvent::Tcp4ApiHandled => { stats_lock.tcp4_api_handled += value; }
-                    StatsEvent::Tcp4AnnouncesHandled => { stats_lock.tcp4_announces_handled += value; }
-                    StatsEvent::Tcp4ScrapesHandled => { stats_lock.tcp4_scrapes_handled += value; }
-                    StatsEvent::Tcp6ConnectionsHandled => { stats_lock.tcp6_connections_handled += value; }
-                    StatsEvent::Tcp6ApiHandled => { stats_lock.tcp6_api_handled += value; }
-                    StatsEvent::Tcp6AnnouncesHandled => { stats_lock.tcp6_announces_handled += value; }
-                    StatsEvent::Tcp6ScrapesHandled => { stats_lock.tcp6_scrapes_handled += value; }
-                    StatsEvent::Udp4ConnectionsHandled => { stats_lock.udp4_connections_handled += value; }
-                    StatsEvent::Udp4AnnouncesHandled => { stats_lock.udp4_announces_handled += value; }
-                    StatsEvent::Udp4ScrapesHandled => { stats_lock.udp4_scrapes_handled += value; }
-                    StatsEvent::Udp6ConnectionsHandled => { stats_lock.udp6_connections_handled += value; }
-                    StatsEvent::Udp6AnnouncesHandled => { stats_lock.udp6_announces_handled += value; }
-                    StatsEvent::Udp6ScrapesHandled => { stats_lock.udp6_scrapes_handled += value; }
-                }
-                let stats = stats_lock.clone();
-                drop(stats_lock);
-                stats
-            }).await {
-                Ok(stats) => { return stats; }
-                Err(_) => {
-                    if count == 5 { panic!("[UPDATE_STATS] Write Lock (stats) request timed out, giving up..."); }
-                    error!("[UPDATE_STATS] Write Lock (stats) request timed out, retrying {} time(s)...", count);
-                    count += 1;
-                }
-            }
-        }
-    }
-
-    pub async fn set_stats(&self, event: StatsEvent, value: i64) -> Stats
-    {
-        let mut count = 1;
-        loop {
-            match timeout(Duration::from_secs(30), async move {
-                let stats_arc = self.stats.clone();
-                let mut stats_lock = stats_arc.lock().await;
-                match event {
-                    StatsEvent::Torrents => { stats_lock.torrents = value; }
-                    StatsEvent::TorrentsUpdates => { stats_lock.torrents_updates = value; }
-                    StatsEvent::TorrentsShadow => { stats_lock.torrents_shadow = value; }
-                    StatsEvent::Users => { stats_lock.users = value; }
-                    StatsEvent::UsersUpdates => { stats_lock.users_updates = value; }
-                    StatsEvent::UsersShadow => { stats_lock.users_shadow = value; }
-                    StatsEvent::TimestampSave => { stats_lock.timestamp_run_save = value; }
-                    StatsEvent::TimestampTimeout => { stats_lock.timestamp_run_timeout = value; }
-                    StatsEvent::TimestampConsole => { stats_lock.timestamp_run_console = value; }
-                    StatsEvent::TimestampKeysTimeout => { stats_lock.timestamp_run_keys_timeout = value; }
-                    StatsEvent::MaintenanceMode => { stats_lock.maintenance_mode = value; }
-                    StatsEvent::Seeds => { stats_lock.seeds = value; }
-                    StatsEvent::Peers => { stats_lock.peers = value; }
-                    StatsEvent::Completed => { stats_lock.completed = value; }
-                    StatsEvent::Whitelist => { stats_lock.whitelist = value; }
-                    StatsEvent::Blacklist => { stats_lock.blacklist = value; }
-                    StatsEvent::Key => { stats_lock.keys = value; }
-                    StatsEvent::Tcp4ConnectionsHandled => { stats_lock.tcp4_connections_handled = value; }
-                    StatsEvent::Tcp4ApiHandled => { stats_lock.tcp4_api_handled = value; }
-                    StatsEvent::Tcp4AnnouncesHandled => { stats_lock.tcp4_announces_handled = value; }
-                    StatsEvent::Tcp4ScrapesHandled => { stats_lock.tcp4_scrapes_handled = value; }
-                    StatsEvent::Tcp6ConnectionsHandled => { stats_lock.tcp6_connections_handled = value; }
-                    StatsEvent::Tcp6ApiHandled => { stats_lock.tcp6_api_handled = value; }
-                    StatsEvent::Tcp6AnnouncesHandled => { stats_lock.tcp6_announces_handled = value; }
-                    StatsEvent::Tcp6ScrapesHandled => { stats_lock.tcp6_scrapes_handled = value; }
-                    StatsEvent::Udp4ConnectionsHandled => { stats_lock.udp4_connections_handled = value; }
-                    StatsEvent::Udp4AnnouncesHandled => { stats_lock.udp4_announces_handled = value; }
-                    StatsEvent::Udp4ScrapesHandled => { stats_lock.udp4_scrapes_handled = value; }
-                    StatsEvent::Udp6ConnectionsHandled => { stats_lock.udp6_connections_handled = value; }
-                    StatsEvent::Udp6AnnouncesHandled => { stats_lock.udp6_announces_handled = value; }
-                    StatsEvent::Udp6ScrapesHandled => { stats_lock.udp6_scrapes_handled = value; }
-                }
-                let stats = stats_lock.clone();
-                drop(stats_lock);
-                stats
-            }).await {
-                Ok(stats) => { return stats; }
-                Err(_) => {
-                    if count == 5 { panic!("[SET_STATS] Write Lock (stats) request timed out, giving up..."); }
-                    error!("[SET_STATS] Write Lock (stats) request timed out, retrying...");
-                    count += 1;
-                }
-            }
-        }
-    }
-
-    /* === Torrents === */
-    pub async fn load_torrents(&self, tracker: Arc<TorrentTracker>)
-    {
-        if let Ok((torrents, completes)) = self.sqlx.load_torrents(tracker.clone()).await {
-            info!("Loaded {} torrents with {} completes.", torrents, completes);
-            self.set_stats(StatsEvent::Completed, completes as i64).await;
-        }
-    }
-
     // pub async fn load_users(&self)
     // {
     //     if let Ok(users) = self.sqlx.load_users().await {
@@ -404,1079 +98,4 @@ impl TorrentTracker {
     //     }
     // }
 
-    pub async fn load_whitelists(&self)
-    {
-        if let Ok(whitelists) = self.sqlx.load_whitelist().await {
-            let mut whitelist_count = 0i64;
-
-            for info_hash in whitelists.iter() {
-                self.add_whitelist(*info_hash, true).await;
-                whitelist_count += 1;
-            }
-
-            info!("Loaded {} whitelists.", whitelist_count);
-        }
-    }
-
-    pub async fn load_blacklists(&self)
-    {
-        if let Ok(blacklists) = self.sqlx.load_blacklist().await {
-            let mut blacklist_count = 0i64;
-
-            for info_hash in blacklists.iter() {
-                self.add_blacklist(*info_hash, true).await;
-                blacklist_count += 1;
-            }
-
-            info!("Loaded {} blacklists.", blacklist_count);
-        }
-    }
-
-    pub async fn load_keys(&self)
-    {
-        if let Ok(keys) = self.sqlx.load_keys().await {
-            let mut keys_count = 0i64;
-
-            for (hash, timeout) in keys.iter() {
-                self.add_key_raw(*hash, *timeout).await;
-                keys_count += 1;
-            }
-
-            info!("Loaded {} keys.", keys_count);
-        }
-    }
-
-    pub async fn save_torrents(&self) -> Result<bool, ()>
-    {
-        if let Ok(shadow) = self.get_shadow().await {
-            if self.sqlx.save_torrents(shadow).await.is_ok() {
-                return Ok(true);
-            }
-            return Ok(false);
-        }
-        Err(())
-    }
-
-    pub async fn save_whitelists(&self) -> bool
-    {
-        let whitelist = self.get_whitelist().await;
-        if self.sqlx.save_whitelist(whitelist.clone()).await.is_ok() {
-            for (info_hash, value) in whitelist.iter() {
-                if value == &0 {
-                    self.remove_whitelist(*info_hash).await;
-                }
-                if value == &2 {
-                    self.add_whitelist(*info_hash, true).await;
-                }
-            }
-            return true;
-        }
-        false
-    }
-
-    pub async fn save_blacklists(&self) -> bool
-    {
-        let blacklist = self.get_blacklist().await;
-        if self.sqlx.save_blacklist(blacklist).await.is_ok() {
-            return true;
-        }
-        false
-    }
-
-    pub async fn save_keys(&self) -> bool
-    {
-        let keys = self.get_keys().await;
-        if self.sqlx.save_keys(keys).await.is_ok() {
-            return true;
-        }
-        false
-    }
-
-    pub async fn add_torrent(&self, info_hash: InfoHash, torrent_entry: TorrentEntryItem, persistent: bool)
-    {
-        let torrents_arc = self.map_torrents.clone();
-        let mut torrents_lock = torrents_arc.lock().await;
-        let _ = torrents_lock.insert(info_hash, torrent_entry.clone());
-        drop(torrents_lock);
-
-        if persistent {
-            self.add_update(
-                info_hash,
-                torrent_entry.completed,
-            ).await;
-        }
-
-        self.update_stats(StatsEvent::Torrents, 1).await;
-    }
-
-    pub async fn add_torrents(&self, torrents: HashMap<InfoHash, TorrentEntryItem>, persistent: bool)
-    {
-        let torrents_arc = self.map_torrents.clone();
-        let mut torrents_lock = torrents_arc.lock().await;
-        let mut updates = HashMap::new();
-        for (info_hash, torrent_entry) in torrents.iter() {
-            let _ = torrents_lock.insert(*info_hash, torrent_entry.clone());
-            updates.insert(*info_hash, torrent_entry.completed);
-        }
-        drop(torrents_lock);
-
-        if persistent {
-            self.add_updates(
-                updates
-            ).await;
-        }
-
-        self.update_stats(StatsEvent::Torrents, torrents.len() as i64).await;
-    }
-
-    pub async fn get_torrent(&self, info_hash: InfoHash) -> Result<Option<TorrentEntry>, ()>
-    {
-        let torrent = match timeout(Duration::from_secs(30), async move {
-            let torrents_arc = self.map_torrents.clone();
-            let torrents_lock = torrents_arc.lock().await;
-            let torrent = torrents_lock.get(&info_hash).cloned();
-            drop(torrents_lock);
-            torrent
-        }).await {
-            Ok(data) => { data }
-            Err(_) => { error!("[GET_TORRENT] Read Lock (torrents) request timed out!"); return Err(()); }
-        };
-
-        let torrent = match torrent {
-            None => { None }
-            Some(data) => {
-                let peers = match timeout(Duration::from_secs(30), async move {
-                    let peers_arc = self.map_peers.clone();
-                    let peers_lock = peers_arc.lock().await;
-                    let peers = match peers_lock.get(&info_hash).cloned() {
-                        None => { BTreeMap::new() }
-                        Some(data) => { data }
-                    };
-                    drop(peers_lock);
-                    peers
-                }).await {
-                    Ok(data) => { data }
-                    Err(_) => { error!("[GET_TORRENT] Read Lock (peers) request timed out!"); return Err(()); }
-                };
-                Some(TorrentEntry {
-                    peers,
-                    completed: data.completed,
-                    seeders: data.seeders,
-                    leechers: data.leechers,
-                })
-            }
-        };
-
-        Ok(torrent)
-    }
-
-    pub async fn get_torrents(&self, hashes: Vec<InfoHash>) -> Result<HashMap<InfoHash, Option<TorrentEntry>>, ()>
-    {
-        let mut return_torrents = HashMap::new();
-
-        for info_hash in hashes.iter() {
-            let torrent = match timeout(Duration::from_secs(30), async move {
-                let torrents_arc = self.map_torrents.clone();
-                let torrents_lock = torrents_arc.lock().await;
-                let torrent = torrents_lock.get(info_hash).cloned();
-                drop(torrents_lock);
-                torrent
-            }).await {
-                Ok(data) => { data }
-                Err(_) => { error!("[GET_TORRENTS] Read Lock (torrents) request timed out!"); return Err(()); }
-            };
-
-            return_torrents.insert(*info_hash, match torrent {
-                None => { None }
-                Some(data) => {
-                    let peers = match timeout(Duration::from_secs(30), async move {
-                        let peers_arc = self.map_peers.clone();
-                        let peers_lock = peers_arc.lock().await;
-                        let peers_data = match peers_lock.get(info_hash).cloned() {
-                            None => { BTreeMap::new() }
-                            Some(data_peers) => { data_peers }
-                        };
-                        drop(peers_lock);
-                        peers_data
-                    }).await {
-                        Ok(data_peers) => { data_peers }
-                        Err(_) => { error!("[GET_TORRENTS] Read Lock (peers) request timed out!"); return Err(()); }
-                    };
-                    Some(TorrentEntry {
-                        peers,
-                        completed: data.completed,
-                        seeders: data.seeders,
-                        leechers: data.leechers,
-                    })
-                }
-            });
-        }
-
-        Ok(return_torrents)
-    }
-
-    pub async fn get_torrents_chunk(&self, skip: u64, amount: u64) -> Result<HashMap<InfoHash, i64>, ()>
-    {
-        match timeout(Duration::from_secs(30), async move {
-            let torrents_arc = self.map_torrents.clone();
-            let torrents_lock = torrents_arc.lock().await;
-            let mut torrents_return: HashMap<InfoHash, i64> = HashMap::new();
-            let mut current_count: u64 = 0;
-            let mut handled_count: u64 = 0;
-            for (info_hash, item) in torrents_lock.iter() {
-                if current_count < skip {
-                    current_count = current_count.add(1);
-                    continue;
-                }
-                if handled_count >= amount {
-                    break;
-                }
-                torrents_return.insert(*info_hash, item.completed);
-                current_count = current_count.add(1);
-                handled_count = handled_count.add(1);
-            }
-            drop(torrents_lock);
-            torrents_return
-        }).await {
-            Ok(data) => { Ok(data) }
-            Err(_) => { Err(()) }
-        }
-    }
-
-    pub async fn remove_torrent(&self, info_hash: InfoHash, persistent: bool)
-    {
-        let mut removed_torrent = false;
-        let mut remove_seeders = 0i64;
-        let mut remove_leechers = 0i64;
-
-        let torrents_arc = self.map_torrents.clone();
-        let mut torrents_lock = torrents_arc.lock().await;
-        let torrent_option = torrents_lock.get(&info_hash);
-        match torrent_option {
-            None => {}
-            Some(data) => {
-                removed_torrent = true;
-                remove_seeders -= data.seeders;
-                remove_leechers -= data.leechers;
-                torrents_lock.remove(&info_hash);
-            }
-        }
-        drop(torrents_lock);
-
-        let peers_arc = self.map_peers.clone();
-        let mut peers_lock = peers_arc.lock().await;
-        peers_lock.remove(&info_hash);
-        drop(peers_lock);
-
-        if persistent {
-            self.remove_update(info_hash).await;
-            self.remove_shadow(info_hash).await;
-        }
-
-        if removed_torrent { self.update_stats(StatsEvent::Torrents, -1).await; }
-        if remove_seeders != 0 { self.update_stats(StatsEvent::Seeds, remove_seeders).await; }
-        if remove_leechers != 0 { self.update_stats(StatsEvent::Peers, remove_leechers).await; }
-    }
-
-    pub async fn remove_torrents(&self, hashes: Vec<InfoHash>, persistent: bool) -> Result<(), ()>
-    {
-        for info_hash in hashes.iter() {
-            let torrent_option = match timeout(Duration::from_secs(30), async move {
-                let torrents_arc = self.map_torrents.clone();
-                let torrents_lock = torrents_arc.lock().await;
-                let torrent_option = torrents_lock.get(info_hash).cloned();
-                drop(torrents_lock);
-                torrent_option
-            }).await {
-                Ok(data) => { data }
-                Err(_) => {
-                    error!("[REMOVE_TORRENTS] Read Lock (torrents) request timed out!");
-                    return Err(());
-                }
-            };
-
-            if torrent_option.is_some() {
-                self.remove_torrent(*info_hash, persistent).await;
-            }
-        }
-        Ok(())
-    }
-
-    /* === Peers === */
-    pub async fn add_peer(&self, info_hash: InfoHash, peer_id: PeerId, peer_entry: TorrentPeer, completed: bool, persistent: bool) -> Result<TorrentEntry, ()>
-    {
-        let mut added_seeder = false;
-        let mut added_leecher = false;
-        let mut removed_seeder = false;
-        let mut removed_leecher = false;
-        let mut completed_applied = false;
-
-        let torrent_input = match timeout(Duration::from_secs(30), async move {
-            let torrents_arc = self.map_torrents.clone();
-            let torrents_lock = torrents_arc.lock().await;
-            let torrent_input = torrents_lock.get(&info_hash).cloned();
-            drop(torrents_lock);
-            torrent_input
-        }).await {
-            Ok(data) => { data }
-            Err(_) => {
-                error!("[ADD_PEER] Read Lock (torrents) request timed out!");
-                return Err(());
-            }
-        };
-
-        let torrent = match torrent_input {
-            None => { TorrentEntry::new() }
-            Some(mut data_torrent) => {
-                let peer = match timeout(Duration::from_secs(30), async move {
-                    let peers_arc = self.map_peers.clone();
-                    let peers_lock = peers_arc.lock().await;
-                    let peer = peers_lock.get(&info_hash).cloned();
-                    drop(peers_lock);
-                    peer
-                }).await {
-                    Ok(data) => { data }
-                    Err(_) => {
-                        error!("[ADD_PEER] Read Lock (peers) request timed out!");
-                        return Err(());
-                    }
-                };
-
-                let mut peers = match peer {
-                    None => { BTreeMap::new() }
-                    Some(data_peers) => { data_peers }
-                };
-
-                match peers.get(&peer_id).cloned() {
-                    None => {
-                        if peer_entry.left == NumberOfBytes(0) {
-                            data_torrent.seeders += 1;
-                            added_seeder = true;
-                            if completed {
-                                data_torrent.completed += 1;
-                                completed_applied = true;
-                            }
-                        } else {
-                            data_torrent.leechers += 1;
-                            added_leecher = true;
-                        }
-                        let _ = peers.insert(peer_id, peer_entry);
-                    }
-                    Some(data_peer) => {
-                        if data_peer.left == NumberOfBytes(0) && peer_entry.left != NumberOfBytes(0) {
-                            data_torrent.seeders -= 1;
-                            data_torrent.leechers += 1;
-                            removed_seeder = true;
-                            added_leecher = true;
-                        } else if data_peer.left != NumberOfBytes(0) && peer_entry.left == NumberOfBytes(0) {
-                            data_torrent.seeders += 1;
-                            data_torrent.leechers -= 1;
-                            added_seeder = true;
-                            removed_leecher = true;
-                            if completed {
-                                data_torrent.completed += 1;
-                                completed_applied = true;
-                            }
-                        }
-                        let _ = peers.insert(peer_id, peer_entry);
-                    }
-                };
-
-                let torrents_arc = self.map_torrents.clone();
-                let mut torrents_lock = torrents_arc.lock().await;
-                torrents_lock.insert(info_hash, data_torrent.clone());
-                drop(torrents_lock);
-
-                let peers_arc = self.map_peers.clone();
-                let mut peers_lock = peers_arc.lock().await;
-                peers_lock.insert(info_hash, peers.clone());
-                drop(peers_lock);
-
-                TorrentEntry {
-                    peers,
-                    completed: data_torrent.completed,
-                    seeders: data_torrent.seeders,
-                    leechers: data_torrent.leechers,
-                }
-            }
-        };
-
-        if persistent && completed {
-            self.add_update(
-                info_hash,
-                torrent.completed,
-            ).await;
-        }
-
-        if added_seeder { self.update_stats(StatsEvent::Seeds, 1).await; }
-        if added_leecher { self.update_stats(StatsEvent::Peers, 1).await; }
-        if removed_seeder { self.update_stats(StatsEvent::Seeds, -1).await; }
-        if removed_leecher { self.update_stats(StatsEvent::Peers, -1).await; }
-        if completed_applied { self.update_stats(StatsEvent::Completed, 1).await; }
-
-        Ok(torrent)
-    }
-
-    pub async fn remove_peer(&self, info_hash: InfoHash, peer_id: PeerId, _persistent: bool) -> Result<TorrentEntry, ()>
-    {
-        let mut removed_seeder = false;
-        let mut removed_leecher = false;
-
-        let torrent_input = match timeout(Duration::from_secs(30), async move {
-            let torrents_arc = self.map_torrents.clone();
-            let torrents_lock = torrents_arc.lock().await;
-            let torrent_input = torrents_lock.get(&info_hash).cloned();
-            drop(torrents_lock);
-            torrent_input
-        }).await {
-            Ok(data) => { data }
-            Err(_) => {
-                error!("[REMOVE_PEER] Read Lock (torrents) request timed out!");
-                return Err(());
-            }
-        };
-
-        let torrent = match torrent_input {
-            None => { TorrentEntry::new() }
-            Some(mut data_torrent) => {
-                let peer = match timeout(Duration::from_secs(30), async move {
-                    let peers_arc = self.map_peers.clone();
-                    let peers_lock = peers_arc.lock().await;
-                    let peer = peers_lock.get(&info_hash).cloned();
-                    drop(peers_lock);
-                    peer
-                }).await {
-                    Ok(data) => { data }
-                    Err(_) => {
-                        error!("[REMOVE_PEER] Read Lock (peers) request timed out!");
-                        return Err(());
-                    }
-                };
-
-                let mut peers = match peer {
-                    None => { BTreeMap::new() }
-                    Some(data_peers) => { data_peers }
-                };
-                let peer_option = peers.get(&peer_id);
-                if peer_option.is_some() {
-                    let peer = *peer_option.unwrap();
-                    if peer.left == NumberOfBytes(0) {
-                        peers.remove(&peer_id);
-                        data_torrent.seeders -= 1;
-                        removed_seeder = true;
-                    } else {
-                        peers.remove(&peer_id);
-                        data_torrent.leechers -= 1;
-                        removed_leecher = true;
-                    }
-                }
-
-                let torrents_arc = self.map_torrents.clone();
-                let mut torrents_lock = torrents_arc.lock().await;
-                torrents_lock.insert(info_hash, data_torrent.clone());
-                drop(torrents_lock);
-
-                if peers.is_empty() {
-                    let peers_arc = self.map_peers.clone();
-                    let mut peers_lock = peers_arc.lock().await;
-                    peers_lock.remove(&info_hash);
-                    drop(peers_lock);
-                } else {
-                    let peers_arc = self.map_peers.clone();
-                    let mut peers_lock = peers_arc.lock().await;
-                    peers_lock.insert(info_hash, peers.clone());
-                    drop(peers_lock);
-                }
-
-                TorrentEntry {
-                    peers,
-                    completed: data_torrent.completed,
-                    seeders: data_torrent.seeders,
-                    leechers: data_torrent.leechers,
-                }
-            }
-        };
-
-        if removed_seeder { self.update_stats(StatsEvent::Seeds, -1).await; }
-        if removed_leecher { self.update_stats(StatsEvent::Peers, -1).await; }
-
-        Ok(torrent)
-    }
-
-    pub async fn remove_peers(&self, peers: Vec<(InfoHash, PeerId)>, _persistent: bool) -> Result<Vec<(InfoHash, PeerId)>, ()>
-    {
-        let mut removed_seeder = 0i64;
-        let mut removed_leecher = 0i64;
-        let mut return_torrententries = Vec::new();
-
-        for (info_hash, peer_id) in peers.iter() {
-            let torrent = match timeout(Duration::from_secs(30), async move {
-                let torrents_arc = self.map_torrents.clone();
-                let torrents_lock = torrents_arc.lock().await;
-                let torrent = torrents_lock.get(&info_hash).cloned();
-                drop(torrents_lock);
-                torrent
-            }).await {
-                Ok(data) => { data }
-                Err(_) => {
-                    error!("[REMOVE_PEERS] Read Lock (torrents) request timed out!");
-                    return Err(());
-                }
-            };
-
-            if let Some(mut data_torrent) = torrent {
-                let peer = match timeout(Duration::from_secs(30), async move {
-                    let peers_arc = self.map_peers.clone();
-                    let peers_lock = peers_arc.lock().await;
-                    let peer = peers_lock.get(&info_hash).cloned();
-                    drop(peers_lock);
-                    peer
-                }).await {
-                    Ok(data) => { data }
-                    Err(_) => {
-                        error!("[REMOVE_PEERS] Read Lock (peers) request timed out!");
-                        return Err(());
-                    }
-                };
-
-                let mut peers = match peer {
-                    None => { BTreeMap::new() }
-                    Some(data_peers) => { data_peers }
-                };
-
-                let peer_option = peers.get(peer_id);
-                if peer_option.is_some() {
-                    let peer = *peer_option.unwrap();
-                    if peer.left == NumberOfBytes(0) {
-                        peers.remove(peer_id);
-                        data_torrent.seeders -= 1;
-                        removed_seeder -= 1;
-                    } else {
-                        peers.remove(peer_id);
-                        data_torrent.leechers -= 1;
-                        removed_leecher -= 1;
-                    }
-                    return_torrententries.push((info_hash.clone(), peer_id.clone()));
-                }
-
-                let torrents_arc = self.map_torrents.clone();
-                let mut torrents_lock = torrents_arc.lock().await;
-                torrents_lock.insert(*info_hash, data_torrent.clone());
-                drop(torrents_lock);
-
-                if peers.is_empty() {
-                    let peers_arc = self.map_peers.clone();
-                    let mut peers_lock = peers_arc.lock().await;
-                    peers_lock.remove(info_hash);
-                    drop(peers_lock);
-                } else {
-                    let peers_arc = self.map_peers.clone();
-                    let mut peers_lock = peers_arc.lock().await;
-                    peers_lock.insert(*info_hash, peers.clone());
-                    drop(peers_lock);
-                }
-            };
-        }
-
-        if removed_seeder != 0 { self.update_stats(StatsEvent::Seeds, removed_seeder).await; }
-        if removed_leecher != 0 { self.update_stats(StatsEvent::Peers, removed_leecher).await; }
-
-        Ok(return_torrententries)
-    }
-
-    pub async fn clean_peers(&self, peer_timeout: Duration)
-    {
-        // Cleaning up peers in chunks, to prevent slow behavior.
-        let mut start: usize = 0;
-        let size: usize = self.config.cleanup_chunks.unwrap_or(100000) as usize;
-        let mut removed_peers = 0u64;
-
-        loop {
-            info!("[PEERS] Scanning peers {} to {}", start, (start + size));
-
-            let peers_arc = self.map_peers.clone();
-            let peers_lock = peers_arc.lock().await;
-            let mut torrent_index = vec![];
-            for (info_hash, _) in peers_lock.iter().skip(start) {
-                torrent_index.push(*info_hash);
-                if torrent_index.len() == size {
-                    break;
-                }
-            }
-            drop(peers_lock);
-
-            let mut peers = vec![];
-            let torrents = match self.get_torrents(torrent_index.clone()).await {
-                Ok(data_request) => { data_request }
-                Err(_) => { continue; }
-            };
-            for (info_hash, torrent_entry) in torrents.iter() {
-                if torrent_entry.is_some() {
-                    let torrent = torrent_entry.clone().unwrap().clone();
-                    for (peer_id, torrent_peer) in torrent.peers.iter() {
-                        if torrent_peer.updated.elapsed() > peer_timeout {
-                            peers.push((*info_hash, *peer_id));
-                        }
-                    }
-                } else {
-                    continue;
-                }
-            }
-            if let Ok(data_request) = self.remove_peers(peers.clone(), self.config.clone().persistence).await {
-                removed_peers += data_request.len() as u64;
-            } else {
-                continue;
-            }
-
-            if torrent_index.len() != size {
-                break;
-            }
-
-            start += size;
-        }
-        info!("[PEERS] Removed {} peers", removed_peers);
-    }
-
-    /* === Updates === */
-    pub async fn add_update(&self, info_hash: InfoHash, completed: i64)
-    {
-        let updates_arc = self.updates.clone();
-        let mut updates_lock = updates_arc.lock().await;
-        updates_lock.insert(info_hash, completed);
-        let update_count = updates_lock.len();
-        drop(updates_lock);
-
-        self.set_stats(StatsEvent::TorrentsUpdates, update_count as i64).await;
-    }
-
-    pub async fn add_updates(&self, updates: HashMap<InfoHash, i64>)
-    {
-        let mut update_count = 0;
-
-        for (info_hash, completed) in updates.iter() {
-            let updates_arc = self.updates.clone();
-            let mut updates_lock = updates_arc.lock().await;
-            updates_lock.insert(*info_hash, *completed);
-            update_count = updates_lock.len();
-            drop(updates_lock);
-        }
-
-        self.set_stats(StatsEvent::TorrentsUpdates, update_count as i64).await;
-    }
-
-    pub async fn get_update(&self) -> Result<HashMap<InfoHash, i64>, ()>
-    {
-        let updates = match timeout(Duration::from_secs(30), async move {
-            let updates_arc = self.updates.clone();
-            let updates_lock = updates_arc.lock().await;
-            let updates = updates_lock.clone();
-            drop(updates_lock);
-            updates
-        }).await {
-            Ok(data) => { data }
-            Err(_) => {
-                error!("[GET_UPDATE] Read Lock (updates) request timed out!");
-                return Err(());
-            }
-        };
-
-        Ok(updates)
-    }
-
-    pub async fn remove_update(&self, info_hash: InfoHash)
-    {
-        let updates_arc = self.updates.clone();
-        let mut updates_lock = updates_arc.lock().await;
-        updates_lock.remove(&info_hash);
-        let update_count = updates_lock.len();
-        drop(updates_lock);
-
-        self.set_stats(StatsEvent::TorrentsUpdates, update_count as i64).await;
-    }
-
-    pub async fn remove_updates(&self, hashes: Vec<InfoHash>)
-    {
-        let mut update_count = 0;
-
-        for info_hash in hashes.iter() {
-            let updates_arc = self.updates.clone();
-            let mut updates_lock = updates_arc.lock().await;
-            updates_lock.remove(info_hash);
-            update_count = updates_lock.len();
-            drop(updates_lock);
-        }
-
-        self.set_stats(StatsEvent::TorrentsUpdates, update_count as i64).await;
-    }
-
-    pub async fn transfer_updates_to_shadow(&self)
-    {
-        let updates_arc = self.updates.clone();
-        let mut updates_lock = updates_arc.lock().await;
-        let updates = updates_lock.clone();
-        updates_lock.clear();
-        drop(updates_lock);
-
-        for (info_hash, completed) in updates.iter() {
-            self.add_shadow(*info_hash, *completed).await;
-        }
-
-        self.set_stats(StatsEvent::TorrentsUpdates, 0).await;
-    }
-
-    /* === Shadow === */
-    pub async fn add_shadow(&self, info_hash: InfoHash, completed: i64)
-    {
-        let shadow_arc = self.shadow.clone();
-        let mut shadow_lock = shadow_arc.lock().await;
-        shadow_lock.insert(info_hash, completed);
-        let shadow_count = shadow_lock.len();
-        drop(shadow_lock);
-
-        self.set_stats(StatsEvent::TorrentsShadow, shadow_count as i64).await;
-    }
-
-    pub async fn remove_shadow(&self, info_hash: InfoHash)
-    {
-        let shadow_arc = self.shadow.clone();
-        let mut shadow_lock = shadow_arc.lock().await;
-        shadow_lock.remove(&info_hash);
-        let shadow_count = shadow_lock.len();
-        drop(shadow_lock);
-
-        self.set_stats(StatsEvent::TorrentsShadow, shadow_count as i64).await;
-    }
-
-    pub async fn remove_shadows(&self, hashes: Vec<InfoHash>)
-    {
-        let mut shadow_count = 0;
-
-        for info_hash in hashes.iter() {
-            let shadow_arc = self.shadow.clone();
-            let mut shadow_lock = shadow_arc.lock().await;
-            shadow_lock.remove(info_hash);
-            shadow_count = shadow_lock.len();
-            drop(shadow_lock);
-        }
-
-        self.set_stats(StatsEvent::TorrentsShadow, shadow_count as i64).await;
-    }
-
-    pub async fn get_shadow(&self) -> Result<HashMap<InfoHash, i64>, ()>
-    {
-        let shadow = match timeout(Duration::from_secs(30), async move {
-            let shadow_arc = self.shadow.clone();
-            let shadow_lock = shadow_arc.lock().await;
-            let shadow = shadow_lock.clone();
-            drop(shadow_lock);
-            shadow
-        }).await {
-            Ok(data) => { data }
-            Err(_) => {
-                error!("[GET_SHADOW] Read Lock (shadow) request timed out!");
-                return Err(());
-            }
-        };
-
-        Ok(shadow)
-    }
-
-    pub async fn clear_shadow(&self)
-    {
-        let shadow_arc = self.shadow.clone();
-        let mut shadow_lock = shadow_arc.lock().await;
-        shadow_lock.clear();
-        drop(shadow_lock);
-
-        self.set_stats(StatsEvent::TorrentsShadow, 0).await;
-    }
-
-    /* === Whitelist === */
-    pub async fn add_whitelist(&self, info_hash: InfoHash, on_load: bool)
-    {
-        let whitelist_arc = self.whitelist.clone();
-        let mut whitelist_lock = whitelist_arc.lock().await;
-        if on_load {
-            whitelist_lock.insert(info_hash, 1i64);
-        } else {
-            whitelist_lock.insert(info_hash, 2i64);
-        }
-        drop(whitelist_lock);
-
-        self.update_stats(StatsEvent::Whitelist, 1).await;
-    }
-
-    pub async fn get_whitelist(&self) -> HashMap<InfoHash, i64>
-    {
-        let mut return_list = HashMap::new();
-
-        let whitelist_arc = self.whitelist.clone();
-        let whitelist_lock = whitelist_arc.lock().await;
-        for (info_hash, value) in whitelist_lock.iter() {
-            return_list.insert(*info_hash, *value);
-        }
-        drop(whitelist_lock);
-
-        return_list
-    }
-
-    pub async fn remove_flag_whitelist(&self, info_hash: InfoHash)
-    {
-        let whitelist_arc = self.whitelist.clone();
-        let mut whitelist_lock = whitelist_arc.lock().await;
-        if whitelist_lock.get(&info_hash).is_some() {
-            whitelist_lock.insert(info_hash, 0i64);
-        }
-        let whitelists = whitelist_lock.clone();
-        drop(whitelist_lock);
-
-        let mut whitelist_count = 0i64;
-        for (_, value) in whitelists.iter() {
-            if value == &1i64 {
-                whitelist_count += 1;
-            }
-        }
-
-        self.set_stats(StatsEvent::Whitelist, whitelist_count).await;
-    }
-
-    pub async fn remove_whitelist(&self, info_hash: InfoHash)
-    {
-        let whitelist_arc = self.whitelist.clone();
-        let mut whitelist_lock = whitelist_arc.lock().await;
-        whitelist_lock.remove(&info_hash);
-        let whitelists = whitelist_lock.clone();
-        drop(whitelist_lock);
-
-        let mut whitelist_count = 0i64;
-        for (_, value) in whitelists.iter() {
-            if value == &1 {
-                whitelist_count += 1;
-            }
-        }
-
-        self.set_stats(StatsEvent::Whitelist, whitelist_count).await;
-    }
-
-    pub async fn check_whitelist(&self, info_hash: InfoHash) -> bool
-    {
-        let whitelist_arc = self.whitelist.clone();
-        let whitelist_lock = whitelist_arc.lock().await;
-        let whitelist = whitelist_lock.get(&info_hash).cloned();
-        drop(whitelist_lock);
-
-        if whitelist.is_some() {
-            return true;
-        }
-
-        false
-    }
-
-    pub async fn clear_whitelist(&self)
-    {
-        let whitelist_arc = self.whitelist.clone();
-        let mut whitelist_lock = whitelist_arc.lock().await;
-        whitelist_lock.clear();
-        drop(whitelist_lock);
-
-        self.set_stats(StatsEvent::Whitelist, 0).await;
-    }
-
-    /* === Blacklist === */
-    pub async fn add_blacklist(&self, info_hash: InfoHash, on_load: bool)
-    {
-        let blacklist_arc = self.blacklist.clone();
-        let mut blacklist_lock = blacklist_arc.lock().await;
-        if on_load {
-            blacklist_lock.insert(info_hash, 1i64);
-        } else {
-            blacklist_lock.insert(info_hash, 2i64);
-        }
-        drop(blacklist_lock);
-
-        self.update_stats(StatsEvent::Blacklist, 1).await;
-    }
-
-    pub async fn get_blacklist(&self) -> Vec<InfoHash>
-    {
-        let mut return_list = vec![];
-
-        let blacklist_arc = self.blacklist.clone();
-        let blacklist_lock = blacklist_arc.lock().await;
-        for (info_hash, _) in blacklist_lock.iter() {
-            return_list.push(*info_hash);
-        }
-        drop(blacklist_lock);
-
-        return_list
-    }
-
-    pub async fn remove_flag_blacklist(&self, info_hash: InfoHash)
-    {
-        let blacklist_arc = self.blacklist.clone();
-        let mut blacklist_lock = blacklist_arc.lock().await;
-        if blacklist_lock.get(&info_hash).is_some() {
-            blacklist_lock.insert(info_hash, 0i64);
-        }
-        let blacklists = blacklist_lock.clone();
-        drop(blacklist_lock);
-
-        let mut blacklist_count = 0i64;
-        for (_, value) in blacklists.iter() {
-            if value == &1i64 {
-                blacklist_count += 1;
-            }
-        }
-
-        self.set_stats(StatsEvent::Blacklist, blacklist_count).await;
-    }
-
-    pub async fn remove_blacklist(&self, info_hash: InfoHash)
-    {
-        let blacklist_arc = self.blacklist.clone();
-        let mut blacklist_lock = blacklist_arc.lock().await;
-        blacklist_lock.remove(&info_hash);
-        let blacklists = blacklist_lock.clone();
-        drop(blacklist_lock);
-
-        let mut blacklist_count = 0i64;
-        for (_, value) in blacklists.iter() {
-            if value == &1 {
-                blacklist_count += 1;
-            }
-        }
-
-        self.set_stats(StatsEvent::Blacklist, blacklist_count).await;
-    }
-
-    pub async fn check_blacklist(&self, info_hash: InfoHash) -> bool
-    {
-        let blacklist_arc = self.blacklist.clone();
-        let blacklist_lock = blacklist_arc.lock().await;
-        let blacklist = blacklist_lock.get(&info_hash).cloned();
-        drop(blacklist_lock);
-
-        if blacklist.is_some() {
-            return true;
-        }
-
-        false
-    }
-
-    pub async fn clear_blacklist(&self)
-    {
-        let blacklist_arc = self.blacklist.clone();
-        let mut blacklist_lock = blacklist_arc.lock().await;
-        blacklist_lock.clear();
-        drop(blacklist_lock);
-
-        self.set_stats(StatsEvent::Blacklist, 0).await;
-    }
-
-    /* === Keys === */
-    pub async fn add_key(&self, hash: InfoHash, timeout: i64)
-    {
-        let keys_arc = self.keys.clone();
-        let mut keys_lock = keys_arc.lock().await;
-        let timestamp = SystemTime::now().duration_since(UNIX_EPOCH).unwrap();
-        let timeout_unix = timestamp.as_secs() as i64 + timeout;
-        keys_lock.insert(hash, timeout_unix);
-        drop(keys_lock);
-
-        self.update_stats(StatsEvent::Key, 1).await;
-    }
-
-    pub async fn add_key_raw(&self, hash: InfoHash, timeout: i64)
-    {
-        let keys_arc = self.keys.clone();
-        let mut keys_lock = keys_arc.lock().await;
-        let time = SystemTime::from(Utc.timestamp_opt(timeout, 0).unwrap());
-        match time.duration_since(SystemTime::now()) {
-            Ok(_) => {
-                keys_lock.insert(hash, timeout);
-            }
-            Err(_) => {
-                drop(keys_lock);
-                return;
-            }
-        }
-        drop(keys_lock);
-
-        self.update_stats(StatsEvent::Key, 1).await;
-    }
-
-    pub async fn get_keys(&self) -> Vec<(InfoHash, i64)>
-    {
-        let keys_arc = self.keys.clone();
-        let keys_lock = keys_arc.lock().await;
-        let keys = keys_lock.clone();
-        drop(keys_lock);
-
-        let mut return_list = vec![];
-        for (hash, timeout) in keys.iter() {
-            return_list.push((*hash, *timeout));
-        }
-
-        return_list
-    }
-
-    pub async fn remove_key(&self, hash: InfoHash)
-    {
-        let keys_arc = self.keys.clone();
-        let mut keys_lock = keys_arc.lock().await;
-        keys_lock.remove(&hash);
-        let key_count = keys_lock.len();
-        drop(keys_lock);
-
-        self.set_stats(StatsEvent::Key, key_count as i64).await;
-    }
-
-    pub async fn check_key(&self, hash: InfoHash) -> bool
-    {
-        let keys_arc = self.keys.clone();
-        let keys_lock = keys_arc.lock().await;
-        let key = keys_lock.get(&hash).cloned();
-        drop(keys_lock);
-
-        if key.is_some() {
-            return true;
-        }
-
-        false
-    }
-
-    pub async fn clear_keys(&self)
-    {
-        let keys_arc = self.keys.clone();
-        let mut keys_lock = keys_arc.lock().await;
-        keys_lock.clear();
-        drop(keys_lock);
-
-        self.set_stats(StatsEvent::Key, 0).await;
-    }
-
-    pub async fn clean_keys(&self)
-    {
-        let keys_arc = self.keys.clone();
-        let keys_lock = keys_arc.lock().await;
-        let keys = keys_lock.clone();
-        drop(keys_lock);
-
-        let mut keys_index = vec![];
-        for (hash, timeout) in keys.iter() {
-            keys_index.push((*hash, *timeout));
-        }
-
-        for (hash, timeout) in keys_index.iter() {
-            if *timeout != 0 {
-                let time = SystemTime::from(Utc.timestamp_opt(*timeout, 0).unwrap());
-                match time.duration_since(SystemTime::now()) {
-                    Ok(_) => {}
-                    Err(_) => {
-                        self.remove_key(*hash).await;
-                    }
-                }
-            }
-        }
-    }
 }
